NICE TEMPLATE

+++ db/create.sql
@@ -59,7 +59,7 @@ CREATE TABLE comics (
   page_count INTEGER,
   publication_date DATE,
   image VARCHAR(255),
-  price INTEGER,
+  price DOUBLE,
   publisher_id INTEGER NOT NULL,
   FOREIGN KEY (publisher_id) REFERENCES publishers(id)
 );
